% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CheckPackage}
\alias{CheckPackage}
\title{Check if the packaegs required installed, if not install them
Load the packages required}
\usage{
CheckPackage(ReqPackage)
}
\arguments{
\item{ReqPackage}{A sequence of names of required packages}
}
\description{
Check if the packaegs required installed, if not install them
Load the packages required
}
\examples{
req_packages <- c("MASS")
CheckPackage(req_packages)
}
