% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{EMRegression}
\alias{EMRegression}
\title{EM regression: the EM algorithm for fitting
Estimate the coefficients for ZCAPM for given response and covariates}
\usage{
EMRegression(
  port.exc.ret,
  mkt.exc.ret,
  mkt.sigma,
  tol,
  MaxIter,
  criterion,
  start_timestamp,
  end_timestamp
)
}
\arguments{
\item{port.exc.ret}{A vector of portfolio excess return}

\item{mkt.exc.ret}{A vector of market excess return}

\item{mkt.sigma}{A vector of market sigma}

\item{tol}{A numeric number specifying tolerance for relative difference}

\item{MaxIter}{A numeric number specifying the largest iteration number}

\item{criterion}{Stopping criterion with 1 for absolute difference and 2 for relative difference}

\item{start_timestamp}{A numeric number indicating the start time point}

\item{end_timestamp}{A numeric number indicating the end time point}

\item{month.idx}{A numeric number indicating current month}
}
\value{
A vector of coefficients' estimates
}
\description{
EM regression: the EM algorithm for fitting
Estimate the coefficients for ZCAPM for given response and covariates
}
\examples{
test.yyyymm <- c(rep(202001, 31), rep(202002, 29), rep(202003, 31))
month.list.test <- unique(test.yyyymm)
month.list.test <- month.list.test[-length(month.list.test)]
y <- runif(length(test.yyyymm), 0, 1)
test.width <- 2
data.factors <- matrix(runif(length(test.yyyymm) * 2, 0, 1), length(test.yyyymm), 2)
tol <- 1e-6
MaxIter <- 100
criterion <- 1
idx <- test.yyyymm \%in\% month.list.test[1]
EMRegression(month.list.test[1], y[idx], data.factors[idx, 1], data.factor[idx, 2], tol, MaxIter, criterion)
}
