% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estZCAPM}
\alias{estZCAPM}
\title{Estimation for ZCAPM}
\usage{
estZCAPM(month.list, yyyymm, port.exc.ret, factors, params)
}
\arguments{
\item{month.list}{A vector of complete month for the observations}

\item{yyyymm}{The year and month of the data to be selected}

\item{port.exc.ret}{A vector of portfolio excess return}

\item{factors, }{the 1st column for mkt.exc.ret, a vector of market excess return,
and the 2nd column for mkt.sigma A vector of market sigma}

\item{params}{A vector (tol, MaxIter, criterion) where
* tol: A numeric number specifying the tolerance of the relative difference of estimates,
* MaxIter: A numeric number specifying the maximum iteration number
* criterion: criterion chosen for the stopping. 1 for absolute difference and 2 for relative difference}
}
\value{
A vector of estimates for coefficients
}
\description{
This function select a period and then get the corresponding response and covariates to
get the ZCAPM estimates using EM.
}
\examples{
test.yyyymm <- c(rep(202001, 31), rep(202002, 29), rep(202003, 31))
month.list.test <- unique(test.yyyymm)
month.list.test <- month.list.test[-length(month.list.test)]
y <- runif(length(test.yyyymm), 0, 1)
test.width <- 2
data.factors <- matrix(runif(length(test.yyyymm) * 2, 0, 1), length(test.yyyymm), 2)
tol <- 1e-6
MaxIter <- 100
criterion <- 1
params <- c(tol, MaxIter, criterion)
estZCAPM(month.list.test[1], test.yyyymm, y, data.factors, params)
}
