% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{FactorModel}
\alias{FactorModel}
\title{Month rolling estimation for linear model}
\usage{
FactorModel(
  port.exc.ret,
  month.list,
  roll.width,
  yyyymm,
  factors,
  intercept_include
)
}
\arguments{
\item{port.exc.ret}{A matrix of portfolios excess return, n_time x n_portfolio}

\item{month.list}{A vector of complete month for the observations}

\item{roll.width}{The width of rolling}

\item{yyyymm}{The year and month of the data to be selected}

\item{factors}{A vector of market excess return}

\item{intercept_include}{bool, whether to return intercept or not}
}
\value{
A list of factor loading and variable names
}
\description{
Month rolling estimation for linear model
}
\examples{
test.yyyymm <- c(rep(202001, 31), rep(202002, 29), rep(202003, 31))
month.list.test <- unique(test.yyyymm)
month.list.test <- month.list.test[-length(month.list.test)]
n_portfolios <- 5
y <- matrix(runif(length(test.yyyymm) * n_portfolios, 0, 1), length(test.yyyymm), n_portfolios)
test.width <- 2
data.factors <- matrix(runif(length(test.yyyymm) * 2, 0, 1), length(test.yyyymm), 2)
colnames(data.factors) <- c("p1", "p2")
intercept_include <- FALSE
FactorModel(y, month.list.test, test.width, test.yyyymm, data.factors, intercept_include)
}
