% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SizedBM25-data.R
\docType{data}
\name{ff25_day}
\alias{ff25_day}
\title{25 Portfolios Formed on Sized Book-to-Market (1960/01/04 - 2018/12/31)}
\format{
An object of class
}
\source{
\href{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/tw_5_ports.html}{Kenneth R. French}
}
\usage{
data(SizedBM25)
}
\description{
This dataset contains value-weighted returns for the intersections of 5 portfolios
formed on size (market equity, ME) and 5 portfolios formed on the ratio of book equity
to market equity (BE/ME). The size breakpoints for year t are the NYSE market equity
quintiles at the end of June of t. BE/ME for June of year t is the book equity for
the last fiscal year end in t-1 divided by ME for December of t-1. The BE/ME
breakpoints are NYSE quintiles.Missing data are indicated by -99.99 or -999.
This dataset includes data from 1960/01/04 to 2018/12/31.
}
\keyword{datasets}
