% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff-factors-data.R
\docType{data}
\name{ff_factors}
\alias{ff_factors}
\title{Five factors, daily(1960/01/04 - 2018/12/31)}
\format{
An object of class
}
\usage{
data(ff_factors)
}
\description{
This dataset includes data from 1960/01/04 to 2018/12/31.
Include Mkt-RF, SMB, HML, MOM and RF.
}
\keyword{datasets}
