% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fm_test.R
\name{fm_test}
\alias{fm_test}
\title{Function to do Fama-MacBeth cross-sectional tests using monthly
rolling approach and single regression approach}
\usage{
fm_test(port.exc.ret.month, fact.load.lst, subsample = NULL)
}
\arguments{
\item{port.exc.ret.month}{one-month ahead monthly return for each portfolio}

\item{fact.load.lst}{a list of factor loadings for each portfolio for each type of models}

\item{subsample}{indicators of which out-of-sample period should be used; default is to use all periods}
}
\value{
a data frame in the form of each panel in Table 2 in the paper
}
\description{
Function to do Fama-MacBeth cross-sectional tests using monthly
rolling approach and single regression approach
}
\examples{
y <- matrix(rnorm(50, 0, 1), 10, 5)
fact.load.1 <- list(factor.loading = matrix(rnorm(100, 0, 1), 20, 5),
                    names.variable = c("p1", "p2"))

fact.load.lst <- list(Model1 = fact.load.1)
test_res <- fm_test(y, fact.load.lst)

}
