% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ind47-data.R
\docType{data}
\name{ind47}
\alias{ind47}
\title{47 Portfolios Formed on 47 Industries (1960/01/04 - 2018/12/31)}
\format{
An object of class
}
\source{
\href{https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/Data_Library/det_48_ind_port.html}{Kenneth R. French}
}
\usage{
data(Ind47)
}
\description{
This dataset contains value- and equal-weighted returns for 47 industry portfolios.
}
\details{
The 47 industries included are agriculture, food products, candy & soda, beer & liquor,
Tobacco products, recreation, entertainment, printing and publishing, consumer goods,
apparel, medical equipment, pharmaceutical products, chemicals, rubber and plastic products,
textiles, construction materials, construction, stell works etc, fabricated products,
machinery, electrical equipment, automobiles and trucks, aircraft, shipbuilding and
railroad equipment, defense, precious metals, non-matallic and industrial metal mining,
coal, petroleum and natrual gas, utilities, communication, personal services, business
services, electronic equipment, measuring and control equipment, business supplies,
shipping containers, transportation, wholesale, retail, restaurants and hotels and motels,
banking, insurance, real estate, trading and other.

The portfolios in the complete file are constructed at the end of June.
Missing data are indicated by -99.99 or -999.
This dataset includes data from 1960/01/04 to 2018/12/31.
}
\keyword{datasets}
