% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MonthRoll}
\alias{MonthRoll}
\title{Monthly Rolling Approach for testing}
\usage{
MonthRoll(idx, port.exc.ret, fact.load.stack)
}
\arguments{
\item{idx}{An indicator vector selecting the desired data}

\item{port.exc.ret}{A vector of portfolio excess return}

\item{fact.load.stack}{A matrix of factor loadings}
}
\value{
Coefficient Estimates for FM testing
}
\description{
Monthly Rolling Approach for testing
}
\examples{
n_timestamps <- 10
n_portfolios <- 5
n_params <- 2
idx <- 1:n_timestamps
y <- matrix(runif(n_timestamps * n_portfolios, 0, 1), n_timestamps, n_portfolios)
fact.load <- matrix(runif(n_timestamps * n_portfolios * n_params, 0, 1), n_timestamps * n_params, n_portfolios)
# select the first timestamp
MonthRoll(idx[1], y, fact.load)
}
