% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu-sigma-data.R
\docType{data}
\name{mu_sigma}
\alias{mu_sigma}
\title{Mu-Sigma factors (1960/01/05 - 2018/12/31)}
\format{
An object of class
}
\usage{
data(mu_sigma)
}
\description{
This dataset contains
This dataset includes data from 1960/01/05 to 2018/12/31.
}
\keyword{datasets}
