% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.summary}
\alias{print.summary}
\title{Print summary for different models}
\usage{
\method{print}{summary}(object, digits = max(4L, getOption("digits")), ...)
}
\arguments{
\item{object}{list of test result, including coefficients and t-values table, R-squared, model names, and variable names}

\item{digits}{digits to display in the table}
}
\value{
Print the summary table
}
\description{
Print summary for different models
}
\examples{
## run fm_test and get the result
y <- matrix(rnorm(50, 0, 1), 10, 5)
fact.load.1 <- list(factor.loading = matrix(rnorm(100, 0, 1), 20, 5),
                    names.variable = c("p1", "p2"))

fact.load.lst <- list(Model1 = fact.load.1)
test_res <- fm_test(y, fact.load.lst)
print.summary(test_res)

## already got the list object with saved results
n_model <- 1
Summary.tab <- vector("list", n_model)
model.names <- vector("list", n_model)
variables <- vector("list", n_model)
Summary.tab[[1]] <- matrix(1:6, 3, 2)
colnames(Summary.tab[[1]]) <- c("coefficient", "t.value")
row.names(Summary.tab[[1]]) <- c("Intercept", "p1", "p2")
r2 <- runif(1, 0, 1)
model.names[[1]] <- "Model1"
variables[[1]] <- c("p1", "p2")

models <- list(Summary.models = Summary.tab,
               r2 = r2,
               model.names = model.names,
               variables = variables)
print.summary(models)
}
