% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RollForward}
\alias{RollForward}
\title{Function to roll forward estimation for one portfolio
return: a matrix whose i-th column is i-th factor loadings}
\usage{
RollForward(port.exc.ret, month.list, width, yyyymm, func, factors, params)
}
\arguments{
\item{port.exc.ret}{A vector of portfolio excess return}

\item{month.list}{A vector of month and year for the complete data}

\item{width}{A numeric specifying the width of the rolling window}

\item{yyyymm}{The selected year and month}

\item{func}{The function to be rolled}

\item{factors}{A matrix of covariates}
}
\value{
Estimates got by rolling forward function for a portfolio
}
\description{
Function to roll forward estimation for one portfolio
return: a matrix whose i-th column is i-th factor loadings
}
\examples{
test.yyyymm <- c(rep(202001, 31), rep(202002, 29), rep(202003, 31))
month.list.test <- unique(test.yyyymm)
month.list.test <- month.list.test[-length(month.list.test)]
y <- runif(length(test.yyyymm), 0, 1)
test.width <- 2
data.factors <- matrix(runif(length(test.yyyymm) * 2, 0, 1), length(test.yyyymm), 2)
include_intercept <- FALSE
RollForward(y, month.list.test, test.width, test.yyyymm, estLM, data.factors, include_intercept)
}
