% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SingleReg}
\alias{SingleReg}
\title{Single Regression Approach for testing}
\usage{
SingleReg(port.exc.ret.avg, fact.load, num.mod, idx.month)
}
\arguments{
\item{port.exc.ret.avg}{A vector of average excess portfolio excess return}

\item{fact.load}{A matrix of the factor loadings}

\item{num.mod}{A numeric specifying the number of models}

\item{idx.month}{An indicator vector selecting the month of interest}
}
\value{
The R2 of the linear regression
}
\description{
Single Regression Approach for testing
}
\examples{
y <- rnorm(5, 0, 1)
names(y) <- paste("stock", 1:5)
num.mod <- 10
num.params <- 2
fact.load <- matrix(rnorm(100, 0, 1), num.mod * num.params, 5)
SingleReg(y, fact.load, num.mod, 1:num.mod)

}
