% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{ZCAPM}
\alias{ZCAPM}
\title{Month rolling estimation for ZCAPM}
\usage{
ZCAPM(port.exc.ret, month.list, roll.width, yyyymm, factors, params)
}
\arguments{
\item{port.exc.ret}{A matrix of portfolios excess return, n_time x n_portfolio}

\item{month.list}{A vector of complete month for the observations}

\item{roll.width}{The width of rolling}

\item{yyyymm}{The year and month of the data to be selected}

\item{factors}{A vector of market excess return}

\item{params}{A vector of tolerance, maximum number of iterations, and stopping criterion for EM
1 for relative difference, `abs(diff(new - old)/old)` and 2 for `abs(diff(new - old)/(abs(old) + 1))`}
}
\value{
A list of factor loading and variable names.
        The factor loading matrix is of size (dim(factors)[2] * n_models) x n_portfolios.
        The variable names contains the names included in the model.
}
\description{
Month rolling estimation for ZCAPM
}
\examples{
test.yyyymm <- c(rep(202001, 31), rep(202002, 29), rep(202003, 31))
month.list.test <- unique(test.yyyymm)
month.list.test <- month.list.test[-length(month.list.test)]
n_portfolios <- 5
y <- matrix(runif(length(test.yyyymm) * n_portfolios, 0, 1), length(test.yyyymm), n_portfolios)
test.width <- 2
data.factors <- matrix(runif(length(test.yyyymm) * 2, 0, 1), length(test.yyyymm), 2)
colnames(data.factors) <- c("p1", "p2")
tol <- 1e-6
MaxIter <- 100
criterion <- 1
params <- c(tol, MaxIter, criterion)
ZCAPM(y, month.list.test, test.width, test.yyyymm, data.factors, params)
}
